Persist game data memento instead of game data directly
Introduction of a game data memento is necessary because there are
multiple forms of persisted game data.  In some cases, the complete game
data must be persisted.  In other cases, notably creating a shallow
clone for AI or odds calculation simulations, the game data delegates
must not be persisted.  Therefore, we allow the user to create a memento
for game data that contains only the information that should ultimately
be persisted.  Thus, the memento itself is what is serialized rather
than the game data directly.